+++ lib/ts/recipe/openid/types.ts
@@ -14,53 +14,25 @@
  */
 import OverrideableBuilder from "supertokens-js-override";
 import type { BaseRequest, BaseResponse } from "../../framework";
-import NormalisedURLDomain from "../../normalisedURLDomain";
-import NormalisedURLPath from "../../normalisedURLPath";
-import { RecipeInterface as JWTRecipeInterface, APIInterface as JWTAPIInterface, JsonWebKey } from "../jwt/types";
 import { GeneralErrorResponse, UserContext } from "../../types";
 
 export type TypeInput = {
-    issuer?: string;
-    jwtValiditySeconds?: number;
     override?: {
         functions?: (
             originalImplementation: RecipeInterface,
             builder?: OverrideableBuilder<RecipeInterface>
         ) => RecipeInterface;
         apis?: (originalImplementation: APIInterface, builder?: OverrideableBuilder<APIInterface>) => APIInterface;
-        jwtFeature?: {
-            functions?: (
-                originalImplementation: JWTRecipeInterface,
-                builder?: OverrideableBuilder<JWTRecipeInterface>
-            ) => JWTRecipeInterface;
-            apis?: (
-                originalImplementation: JWTAPIInterface,
-                builder?: OverrideableBuilder<JWTAPIInterface>
-            ) => JWTAPIInterface;
-        };
     };
 };
 
 export type TypeNormalisedInput = {
-    issuerDomain: NormalisedURLDomain;
-    issuerPath: NormalisedURLPath;
-    jwtValiditySeconds?: number;
     override: {
         functions: (
             originalImplementation: RecipeInterface,
             builder?: OverrideableBuilder<RecipeInterface>
         ) => RecipeInterface;
         apis: (originalImplementation: APIInterface, builder?: OverrideableBuilder<APIInterface>) => APIInterface;
-        jwtFeature?: {
-            functions?: (
-                originalImplementation: JWTRecipeInterface,
-                builder?: OverrideableBuilder<JWTRecipeInterface>
-            ) => JWTRecipeInterface;
-            apis?: (
-                originalImplementation: JWTAPIInterface,
-                builder?: OverrideableBuilder<JWTAPIInterface>
-            ) => JWTAPIInterface;
-        };
     };
 };
 
@@ -128,10 +100,4 @@ export type RecipeInterface = {
               status: "UNSUPPORTED_ALGORITHM_ERROR";
           }
     >;
-
-    getJWKS(input: {
-        userContext: UserContext;
-    }): Promise<{
-        keys: JsonWebKey[];
-    }>;
 };