+++ test/openid/config.test.js
@@ -40,10 +40,7 @@ describe(`configTest: ${printPath("[test/openid/config.test.js]")}`, function ()
             return;
         }
 
-        let openIdRecipe = await OpenIdRecipe.getInstanceOrThrowError();
-
-        assert(openIdRecipe.config.issuerDomain.getAsStringDangerous() === "https://api.supertokens.io");
-        assert(openIdRecipe.config.issuerPath.getAsStringDangerous() === "/auth");
+        assert((await OpenIdRecipe.getIssuer()) === "https://api.supertokens.io/auth");
     });
 
     it("Test that the default config sets values correctly for OpenID recipe with apiBasePath", async function () {
@@ -68,10 +65,7 @@ describe(`configTest: ${printPath("[test/openid/config.test.js]")}`, function ()
             return;
         }
 
-        let openIdRecipe = await OpenIdRecipe.getInstanceOrThrowError();
-
-        assert(openIdRecipe.config.issuerDomain.getAsStringDangerous() === "https://api.supertokens.io");
-        assert(openIdRecipe.config.issuerPath.getAsStringDangerous() === "");
+        assert((await OpenIdRecipe.getIssuer()) === "https://api.supertokens.io");
     });
 
     it("Test that the config sets values correctly for OpenID recipe with issuer", async function () {
@@ -88,7 +82,18 @@ describe(`configTest: ${printPath("[test/openid/config.test.js]")}`, function ()
             },
             recipeList: [
                 OpenIdRecipe.init({
-                    issuer: "https://customissuer.com",
+                    override: {
+                        functions: (originalImplementation) => ({
+                            ...originalImplementation,
+                            getOpenIdDiscoveryConfiguration: async (input) => {
+                                const orig = originalImplementation.getOpenIdDiscoveryConfiguration(input);
+                                return {
+                                    ...orig,
+                                    issuer: "https://customissuer.com",
+                                };
+                            },
+                        }),
+                    },
                 }),
             ],
         });
@@ -100,38 +105,7 @@ describe(`configTest: ${printPath("[test/openid/config.test.js]")}`, function ()
             return;
         }
 
-        let openIdRecipe = await OpenIdRecipe.getInstanceOrThrowError();
-
-        assert(openIdRecipe.config.issuerDomain.getAsStringDangerous() === "https://customissuer.com");
-        assert(openIdRecipe.config.issuerPath.getAsStringDangerous() === "");
-    });
-
-    it("Test that issuer without apiBasePath throws error", async function () {
-        const connectionURI = await startST();
-
-        try {
-            STExpress.init({
-                supertokens: {
-                    connectionURI,
-                },
-                appInfo: {
-                    apiDomain: "api.supertokens.io",
-                    appName: "SuperTokens",
-                    websiteDomain: "supertokens.io",
-                },
-                recipeList: [
-                    OpenIdRecipe.init({
-                        issuer: "https://customissuer.com",
-                    }),
-                ],
-            });
-        } catch (e) {
-            if (
-                e.message !== "The path of the issuer URL must be equal to the apiBasePath. The default value is /auth"
-            ) {
-                throw e;
-            }
-        }
+        assert((await OpenIdRecipe.getIssuer()) === "https://customissuer.com");
     });
 
     it("Test that issuer with gateway path works fine", async function () {
@@ -156,9 +130,6 @@ describe(`configTest: ${printPath("[test/openid/config.test.js]")}`, function ()
             return;
         }
 
-        let openIdRecipe = await OpenIdRecipe.getInstanceOrThrowError();
-
-        assert.equal(openIdRecipe.config.issuerDomain.getAsStringDangerous(), "https://api.supertokens.io");
-        assert.equal(openIdRecipe.config.issuerPath.getAsStringDangerous(), "/gateway/auth");
+        assert.equal(await OpenIdRecipe.getIssuer(), "https://api.supertokens.io/gateway/auth");
     });
 });