+++ test/with-typescript/index.ts
@@ -11,6 +11,7 @@ import NextJS from "../../nextjs";
 import ThirdParty from "../../recipe/thirdparty";
 import Multitenancy from "../../recipe/multitenancy";
 import Passwordless from "../../recipe/passwordless";
+import OpenId from "../../recipe/openid";
 import { SMTPService as SMTPServiceTPP } from "../../recipe/passwordless/emaildelivery";
 import { SMTPService as SMTPServiceP } from "../../recipe/passwordless/emaildelivery";
 import { SMTPService as SMTPServiceTPEP } from "../../recipe/emailpassword/emaildelivery";
@@ -1606,25 +1607,28 @@ Session.init({
                 });
             },
         }),
-        openIdFeature: {
-            functions: (oI) => ({
-                ...oI,
-                getOpenIdDiscoveryConfiguration: async (input) => ({
-                    issuer: "your issuer",
-                    jwks_uri: "https://your.api.domain/auth/jwt/jwks.json",
-                    token_endpoint: "http://localhost:3000/auth/oauth2/token",
-                    authorization_endpoint: "http://localhost:3000/auth/oauth2/auth",
-                    userinfo_endpoint: "http://localhost:3000/auth/oauth2/userinfo",
-                    revocation_endpoint: "http://localhost:3000/auth/oauth2/revoke",
-                    token_introspection_endpoint: "http://localhost:3000/auth/oauth2/introspect",
-                    end_session_endpoint: "http://localhost:3000/auth/oauth2/introspect",
-                    id_token_signing_alg_values_supported: [],
-                    response_types_supported: [],
-                    subject_types_supported: [],
-                    status: "OK",
-                }),
+    },
+});
+
+OpenId.init({
+    override: {
+        functions: (oI) => ({
+            ...oI,
+            getOpenIdDiscoveryConfiguration: async (input) => ({
+                issuer: "your issuer",
+                jwks_uri: "https://your.api.domain/auth/jwt/jwks.json",
+                token_endpoint: "http://localhost:3000/auth/oauth2/token",
+                authorization_endpoint: "http://localhost:3000/auth/oauth2/auth",
+                userinfo_endpoint: "http://localhost:3000/auth/oauth2/userinfo",
+                revocation_endpoint: "http://localhost:3000/auth/oauth2/revoke",
+                token_introspection_endpoint: "http://localhost:3000/auth/oauth2/introspect",
+                end_session_endpoint: "http://localhost:3000/auth/oauth2/introspect",
+                id_token_signing_alg_values_supported: [],
+                response_types_supported: [],
+                subject_types_supported: [],
+                status: "OK",
             }),
-        },
+        }),
     },
 });
 
@@ -1886,21 +1890,17 @@ Passwordless.init({
 
 const recipeUserId = new Supertokens.RecipeUserId("asdf");
 
-Session.init({
+JWT.init({
     override: {
-        openIdFeature: {
-            jwtFeature: {
-                apis: (oI) => {
-                    return {
-                        ...oI,
-                        getJWKSGET: async function (input) {
-                            let result = await oI.getJWKSGET!(input);
-                            input.options.res.setHeader("custom-header", "custom-value", false);
-                            return result;
-                        },
-                    };
+        apis: (oI) => {
+            return {
+                ...oI,
+                getJWKSGET: async function (input) {
+                    let result = await oI.getJWKSGET!(input);
+                    input.options.res.setHeader("custom-header", "custom-value", false);
+                    return result;
                 },
-            },
+            };
         },
     },
 });